refactor: 디렉토리 구조 controller/domain/repository/service로 분리
