+++ src/test/java/subway/SubwayLineAcceptanceTest.java
@@ -1,0 +1,2 @@
+package subway;public class SubwayLineAcceptanceTest {
+}
